Refactoring of pagination on resources and other small fixes
-logo's clickable area is now smaller on mobile and does not conflict with hamburger button
-total shows 0 if not results
