+++ components/NewsArticle.jsx
@@ -1,4 +1,3 @@
-import Image from 'next/image';
 import Link from 'next/link';
 import cn from 'classnames';
 import Markdown from 'markdown-to-jsx';
