+++ components/Pagination.jsx
@@ -1,13 +1,28 @@
+import React, { useState, useEffect } from 'react';
+import { useRouter } from 'next/router';
+import { cleanupQuery, formatQuerystring } from '../lib/utils';
 import cn from 'classnames';
 import { usePagination, DOTS } from '../hooks/usePagination';
 import styles from '../styles/Pagination.module.scss';
+
 const Pagination = props => {
+
+  const router = useRouter();
+  const [currentPage, setCurrentPage] = useState(1);
+
+  useEffect(
+    () => {
+      const { query } = router;
+      query.p = currentPage;
+      router.push(router.pathname + formatQuerystring(cleanupQuery(query)), undefined, { scroll: false });
+    }, [currentPage]
+  );
+
   const {
-    onPageChange,
     maxPage,
     siblingCount = 1,
-    currentPage,
-    className
+    className,
+    themeColor
   } = props;
 
   const paginationRange = usePagination({
@@ -22,11 +37,11 @@ const Pagination = props => {
   }
 
   const onNext = () => {
-    onPageChange(currentPage + 1);
+    setCurrentPage(currentPage + 1);
   };
 
   const onPrevious = () => {
-    onPageChange(currentPage - 1);
+    setCurrentPage(currentPage - 1);
   };
 
   let lastPage = paginationRange[paginationRange.length - 1];
@@ -35,7 +50,9 @@ const Pagination = props => {
     <ul className={cn(styles['pagination-container'], { [className]: className })} >
        {/* Left navigation */}
       <li >
-        <div onClick={onPrevious} className={cn(styles['pagination-item'], styles.prev, { [styles.disabled]: currentPage === 1 })}>Previous</div>
+        <div onClick={onPrevious} className={cn(styles['pagination-item'], styles.prev, { [styles.disabled]: currentPage === 1 })}>
+          <i className="ri-arrow-left-line"/>
+        </div>
       </li>
       {paginationRange.map( (pageNumber, i)=> {
          
@@ -46,15 +63,17 @@ const Pagination = props => {
 		
         // Render our Page Pills
         return (
-          <li key={i} onClick={() => onPageChange(pageNumber)}>
-            <div className={cn(styles['pagination-item'], { [styles.selected]: pageNumber === currentPage})}>{pageNumber}</div>
+          <li key={i} onClick={() => setCurrentPage(pageNumber)}>
+            <div className={cn(styles['pagination-item'], { [styles.selected]: pageNumber === currentPage })}>{pageNumber}</div>
           </li>
         );
       })}
 
       {/*  Right Navigation */}
       <li>
-        <div onClick={onNext} className={cn(styles['pagination-item'], styles.next, { [styles.disabled]: currentPage === lastPage })}>Next</div>
+        <div onClick={onNext} className={cn(styles['pagination-item'], styles.next, { [styles.disabled]: currentPage === lastPage })}>
+          <i className="ri-arrow-right-line"/>
+        </div>
       </li>
     </ul>
   );
