+++ hooks/usePagination.js
@@ -1,5 +1,7 @@
 import { useMemo } from 'react';
 
+export const DOTS = '...';
+
 export const usePagination = ({
   maxPage,
   siblingCount = 1,
