+++ lib/api.js
@@ -30,7 +30,6 @@ author {
 }
 attributeOriginalAuthor
 originalAuthor
-tags
 featuredImage {
   url
 }
