+++ styles/Hero.module.scss
@@ -73,6 +73,7 @@
 			font-size: 4rem;
 			line-height: 5rem;
 			max-width: 38rem;
+			color: white;
 		}
 		p {
 			margin: 1.5em 0;
