+++ styles/NewsArticle.module.scss
@@ -36,7 +36,7 @@
 }
 
 .separator {
-    border: 1px solid black;
+    border: 1px solid var(--scp-color-dark-grey);
 }
 
 .landscape {
