+++ styles/NewsFeed.module.scss
@@ -42,6 +42,7 @@
 			color: var(--scp-color-sacramento);
 
 			h2 {
+				color: var(--scp-color-sacramento);
 				margin: 0;
 			}
 		}
@@ -166,7 +167,6 @@
 				}
 			}
 			> .top-paginator {
-				margin-right: 1rem;
 				@media (max-width: 900px) {
 					display: none;
 				}
@@ -182,8 +182,8 @@
 			display: flex;
 			margin-top: 1.4rem;
 			margin-bottom: 1rem;
-			margin-left: 1rem;
 			@media (max-width: 900px) {
+				margin-left: 0;
 				justify-content: center;
 			}
 		}
@@ -210,7 +210,6 @@
 		flex-wrap: wrap;
 		@media (max-width: 1400px) {
 			gap: 1rem;
-			margin-bottom: 1rem;
 			justify-content: center;
 		}
 	}
