+++ styles/NewsPreview.module.scss
@@ -16,7 +16,6 @@
 	}
 
 	.button::before {
-		content: '>>';
 		padding-right: .5em;
 	}
 	.button {
