+++ styles/ResourceEntry.module.scss
@@ -52,7 +52,7 @@
 }
 
 .separator {
-    border: 1px solid black;
+    border: 1px solid var(--scp-color-dark-grey);
 }
 
 .download-section {
