+++ styles/ResourceFeed.module.scss
@@ -43,6 +43,7 @@
 			color: var(--scp-color-eggplant);
 
 			h2 {
+				color: var(--scp-color-eggplant);
 				margin: 0;
 			}
 		}
@@ -79,7 +80,7 @@
 					display: relative;
 					margin: 1.5rem 0 0 0;
 					border: 1px solid var(--scp-color-eggplant);
-					height: 2.3rem;
+					height: 2.5rem;
 					min-width: 14rem;
 					@media (max-width: 900px) {
 						width:auto;
@@ -116,7 +117,7 @@
 				}
 
 				> .category {
-					height: 2.4rem;
+					height: 2.5rem;
 					min-width: 14rem;
 					margin: 0 auto 0 0rem;
 					display: inline-block;
@@ -130,24 +131,10 @@
 						font-family: 'Heebo', sans-serif;
 						font-size: .8em;
 					}
-
-					> .category-list {
-						margin: 0;
-						height: 2.4rem;
-						border: 2px solid var(--scp-color-eggplant);
-						border-radius: 0;
-						font-size: 16px;
-						-webkit-appearance: none;
-						padding-left: .5rem;
-						@media (max-width: 900px) {
-							margin: 0;
-							width: 100%;
-						}
-					}
 				}
 
 				> .subject {
-					height: 2.4rem;
+					height: 2.5rem;
 					min-width: 14rem;
 					display: inline-block;
 					@media (max-width: 900px) {
@@ -160,20 +147,6 @@
 						margin: 0;
 						font-family: 'Heebo', sans-serif;
 						font-size: .8em;
-					}
-
-					> .subject-list {
-						margin: 0;
-						height: 2.4rem;
-						border: 2px solid var(--scp-color-eggplant);
-						border-radius: 0;
-						font-size: 16px;
-						-webkit-appearance: none;
-						padding-left: .5rem;
-						@media (max-width: 900px) {
-							margin: 0;
-							width: 100%;
-						}
 					}
 				}
 				> .total-reset {
@@ -229,6 +202,7 @@
 			}
 
 			> .top-paginator {
+				margin-top: 1.5rem;
 				@media (max-width: 900px) {
 					display: none;
 				}
@@ -294,6 +268,16 @@
 				> .page-item:hover:not(.active):not(.disabled) {
 					background-color: #ddd;
 				}
+			}
+		}
+
+		> .bottom-paginator {
+			display: flex;
+			margin-top: 1.4rem;
+			margin-bottom: 1rem;
+			@media (max-width: 900px) {
+				margin-left: 0;
+				justify-content: center;
 			}
 		}
