+++ styles/Team.module.scss
@@ -119,6 +119,7 @@
 			margin-top: 0.5rem;
 			margin-bottom: 0.25rem;
 			display: flex;
+			color: white;
 	
 			@media (max-width: 900px) {
 				font-size: 1.3rem;
